+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Pre validate client with authorization settings
 
 ### Fixed
+- Calculate import checksum after variable substitution
 
 ## [2.5.0] - 2020-10-19